core/src/main/kotlin/org/eazyportal/plugin/release/core/action/FinalizeReleaseVersionAction.kt
package org.eazyportal.plugin.release.core.action

import org.eazyportal.plugin.release.core.project.model.ProjectDescriptor
import org.eazyportal.plugin.release.core.scm.ScmActions
import org.slf4j.Logger
import org.slf4j.LoggerFactory

class FinalizeReleaseVersionAction(
    private val projectDescriptor: ProjectDescriptor,
    private val scmActions: ScmActions,
) : ReleaseAction {

    override fun execute() {
        LOGGER.info("Finalize release version...")

        val releaseVersion = projectDescriptor.rootProject.projectActions.getVersion()

        projectDescriptor.allProjects.forEach {
            scmActions.add(it.dir, *it.projectActions.scmFilesToCommit())
            scmActions.commit(it.dir, "Release version: $releaseVersion")

            scmActions.tag(it.dir, releaseVersion)
        }
    }

    companion object {
        private val LOGGER: Logger = LoggerFactory.getLogger(FinalizeReleaseVersionAction::class.java)
    }

}
